refactor(runtime-worker): rename 12 P2 violations (#1205)
- entity/store→storeTypes, buildSessionBroadcast→Utils,
  command/registry.types→registryTypes, draft/lookupOperations→Utils,
  draft/nameUtilities→nameUtils, route/ideGsmCsv→Utils,
  routeQueryCacheConfig→Constants, shapeChunkStore→Utils,
  stageAlias→Constants, utils/policy-c→policyCUtils,
  utils/metrics→metricsUtils, utils/workerLogger→workerLoggerUtils

Rollback: git revert <hash> --no-edit
