+++ packages/runtime-worker/src/services/NodeLifecycleManager.ts
@@ -2,7 +2,7 @@ import type { NodeId, NodeType } from '@hierarchidb/core-types';
 import type { TreeNode } from '@hierarchidb/tree-api';
 import { SingletonMixin } from '@hierarchidb/util';
 import type { RuntimePluginDefinition } from '~/types/RuntimePluginDefinition';
-import { workerError } from '~/utils/workerLogger';
+import { workerError } from '~/utils/workerLoggerUtils';
 import type { CoreDB } from './CoreDB.js';
 import type { LifecycleContext, LifecycleEvent, NodeLifecycleHooks } from './lifecycle-types.js';
