+++ packages/runtime-worker/src/services/command/registryTypes.ts
